Only delete the message in the input modifier.
The former behavior prevents messages from appearing while still wishing to use the AI, as it will be deleted when the context modifier is entered.

This does have the side effect of continuing to show the message when the player continues without input (makes the AI write more), but I'm not sure how to detect that condition yet.